refactor voting module
