+++ handler_modules/voting_system/exceptions.py
@@ -1,0 +1,18 @@
+class VotingCompletedError(Exception):
+    pass
+
+
+class CharIDNotFoundError(Exception):
+    pass
+
+
+class NoLegitAnimeError(Exception):
+    pass
+
+
+class AlreadyReleasedError(Exception):
+    pass
+
+
+class MalformedABParamsError(Exception):
+    pass
